Fix EXSWSTRHPC-113: rocRAND generates incorrect values for mrg_uniform_distribution_double() (#187)
* Fix a warning about lossy implicit conversion of UINT64_MAX to double

* Improve a warning about inline asm, do not show it for host

#pragma warning is an unknown pragma, Clang support #warning.

* Fix incorrect ranges of uniform float/double device funcs of MRG32K3A

rocrand_uniform, rocrand_uniform_double (and normal/log-normal
distributions) must use raw engine values in
[1, ROCRAND_MRG32K3A_M1] range, not values already mapped to
[0, UINT_MAX] range.
Add tests for range testing.

Fixes EXSWSTRHPC-113.

* Fix compilation of tests on CUDA


* Enable Sobol64 in rocrand_generate tests

* Update CHANGELOG for ROCm 4.4

+++ test/test_rocrand_kernel_sobol64.cpp
@@ -183,7 +183,7 @@ TEST(rocrand_kernel_sobol64, rocrand)
     double mean = 0;
     for(auto v : output_host)
     {
-        mean += static_cast<double>(v) / UINT64_MAX;
+        mean += static_cast<double>(v) / static_cast<double>(UINT64_MAX);
     }
     mean = mean / output_size;
     EXPECT_NEAR(mean, 0.5, 0.1);
